requirements.txt
ollama
chromadb
PyMuPDF
python-docx
pytesseract
Pillow
streamlit
tqdm
openai-whisper